network/ai_training/requirements.txt
# ===== 深度学习框架 =====
torch>=2.0.0
torchvision>=0.15.0

# ===== 强化学习 =====
gymnasium>=0.29.0
# stable-baselines3>=2.0.0  # 可选：使用现成的RL算法实现

# ===== 数据处理 =====
numpy>=1.24.0
pandas>=2.0.0

# ===== 可视化 =====
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# ===== 工具 =====
tqdm>=4.65.0
pyyaml>=6.0
# wandb>=0.15.0  # 可选：实验跟踪工具

# ===== 游戏依赖（与主项目保持一致）=====
# pygame==2.5.2  # 如果需要可视化，可以取消注释